The first stage of making registers work reliably. the original feature of having the body of a token be the description until description was encountered was a serious problem. The stack was never corrected when a <description> element was encountered. Fixing this serious problem and taking care of a few others involves creating two types on the @registers stack.
ptr     = ARRAY [ element , key, ref ]
default = ARRAY [ recursion-level , ref ]

the default is set every time a new define is encountered. This ensures that there is always a ref
for Register::push and that functions can be bounded to the current define. It is also possible
to catch some forms of stack corruption as well now that there are two types, one of which can
be matched against define pops.

Register::replace_default is a new function that allows a default to be replaced with a specific
element properly.

This commit introduces the changes in Register:: without turning on the new code. It passed
the perl5 regression.


git-svn-id: svn://localhost/mattie-subversion@382 534069c3-6cd1-4c95-a44b-7f597ffc1621