Move the parameters parsed from the command line (except the input file list) to the Settings module so that the use of those values is more descriptive in their use context.
Remove the support for generating multiple languages at a time. This was a foolish idea
as the translater can be externally mapped across a language set without enormously
increasing the complexity in the core code.

This commit is almost twice as many deletes as changes which is good. passed the regression
so it should be okay.


git-svn-id: svn://localhost/mattie-subversion@403 534069c3-6cd1-4c95-a44b-7f597ffc1621