somewhat extensive re-factor.
parser-match-function has replaced both parser-make-match and parser-get-match.
This function is simpler than the previous two, and directly associates a function
with the Match Function type referred to in the docs.

I have created a ADT with inline functions creating a Match Result type. functions
include parser-make-match,parser-match-consumed,parser-match-data,parser-make-match-data

I have replaced parser-next with parser-consumed, removing the +1 calculation, which
moved to parser-advance. The previous form was ugly at the least.


git-svn-id: svn://localhost/mattie-subversion@666 534069c3-6cd1-4c95-a44b-7f597ffc1621