M robust-load-elisp
This function used to load an elisp file with error trapping. It now is a
macro that traps errors from expressions to increase the granularity of
error trapping available for files that configure a diverse range
of packages and features such as user.el.

A robust-load-file

The old robust-load-elisp function renamed more appropriately.

Some better documentation of how load-path is formed has been added as well.


git-svn-id: svn://localhost/mattie-subversion@1172 534069c3-6cd1-4c95-a44b-7f597ffc1621