start the transition to a new yasnippet based template system.
* else-xml,xml-code
else-xml is renamed to xml-code and deprecated.

A small set of stubs, and a preliminary sketch of the library is written up.

The old xml code is commented out until the perl based template assembler is
converted to the yasnippet syntax.

* programming.el

this file was modified to pull out all of the old else-xml configuration
and use the xml-code stubs to minimize the trauma of the transition.

* complete.el

make the icicles more robust since it is broken regularly by emacs CVS

* template.el

transition to xml-code.


git-svn-id: svn://localhost/mattie-subversion@1199 534069c3-6cd1-4c95-a44b-7f597ffc1621