the error handling in grail was broken, because errors in loading grail itself were not trapped.
The goal of these changes is to properly detect when grail is not enabled or installed correcly.
When crippled it should trap errors, and allow the normal emacs startup to proceed, with error
messages in the message and scratch buffers.

The error handling has been consolidated into grail-dup-error-to-scratch. Other than defining basic functions the entire grail startup is protected by a condition-case.


git-svn-id: svn://localhost/mattie-subversion@1293 534069c3-6cd1-4c95-a44b-7f597ffc1621