Significant re-work of grail, and my configuration.
Grail was seriously broken by emacs --daemon since window-system
would return false as there was no frame created yet.

Getting the GUI configuration to work correctly required placing
the loading of gui.el on after-make-frame-functions.

The custom-theme-set-faces is totally broken for reasons unkown
to me so I bypassed it with a new grail-set-faces macro.

the call chain for loading user elisp has been cleaned up,
it was previously a disgusting mess.

There is still quite a bit of breakage, especially in spell
that has not been tracked down, but enough is working now
to commit a snapshot.


git-svn-id: svn://localhost/mattie-subversion@1409 534069c3-6cd1-4c95-a44b-7f597ffc1621