----------------------------------------------------------------------
commit-file multi-branch support.

The accummulation of changes into task branches means that the drafting
of the commit log should be relative to a branch. This de-emphasises the
changelogs of individual commits, though they are still important.

This also allows the commit draft to be a planning document organizing
the development of a task branch which is a powerful change to the
workflow.

file: commit-file.el
  [edit-commit-draft]
    * edit commit files on more than one branch

    * rename the buffer to branch:commit so that
      it is clear which branch the commit draft applies to.

  [delete-commit-draft]
    * update delete commit draft to prompt for a branch as well.

  [logedit-insert-commit-changelog-files]
    FIX: add a missing test in the insert loop to ensure that
         a commit file is not inserted twice.

----------------------------------------------------------------------

Create a new library cm-path.el to collect path manipulation functions.

added: cm-path.el

modified: cm-util.el
  * move file-path-if-readable and delete-trailing-path-separators to the new cm-path library.

----------------------------------------------------------------------

added: local/elisp/rc-navigate.el

Create a rc-navigate.el with the branch prompting functions from
bzr-pick.el. The branch prompting now has a sane looking prompt now
that I dosvered the trailing : magic in read-file-name.

modified: bzr-pick.el
  * load cm-path as a dependency instead of cm-string and cm-util

[bzr-prompt-for-branch-dir]
  * update the documentation warning against using this function
    directly, bzr-prompt-for-branch should be used instead.

[bzr-prompt-for-branch]
  * ensure there is a trailing separator and sanitize the path.




git-svn-id: svn://localhost/mattie-subversion@1501 534069c3-6cd1-4c95-a44b-7f597ffc1621