A large cleanup of experiments/alpha.el
originally alpha.el was meant to be a very stable set of commands that needed
just a bit more tweaking before going into commands.el. in fact a better name
would be beta, not alpha.

It grew for some time with all kinds of random experiments being dumped in it.

This mucks out alpha.el moving the really experimental/old stuff elsewhere or
deleting it entirely.



git-svn-id: svn://localhost/mattie-subversion@1518 534069c3-6cd1-4c95-a44b-7f597ffc1621