Bump grail to 2.0
Grail now supports orders in loading groups. This allows
the groups to be loaded in a specified order so that
groups can be dependent, or use the effects of other groups.

The order is a simple >= 0 integer, loaded from lowest to
highest.

[use-grail-groups]
  * add an order parameter.

[grail-load-requested-groups]
  * sort the requested groups by their order

programming.el and user.el are migrated to the new
use-grail-groups API.




git-svn-id: svn://localhost/mattie-subversion@1536 534069c3-6cd1-4c95-a44b-7f597ffc1621