* Switch over from grail-groups to grail-profile for installation suppport. * ELPA is removed entirely. * support for installing from checkouts of several VC systems.
This changeset has two goals: simplifying to the common case, and removing stuff that doesn't work.

The original grail-groups design with the grail-activate-with-recovery macro was a more complete
solution to capturing the initialization code, and being able to defer execution of it until
the loading of dependencies was assured.

The massive cleanup of grail-groups from the strange and hairy codebase to something clean was
moved over to the profile concept. Instead of capturing the initialization code a new function
grail-load was created that throws 'grail-trap if loading and installation both fail to provide
a needed dependency.

Basically all the init code can be placed after putting grail-load calls at the top of the file.

ELPA was removed because the suppport for it was poor.



git-svn-id: svn://localhost/mattie-subversion@1573 534069c3-6cd1-4c95-a44b-7f597ffc1621