summary: fix and complete the completion system
description:

the completion system is quite complicated due to the fact that
I am trying to reconcile and use enhanced completion. construct
a standard interface to a back-end flexible completion system
to use

[local/profiles/dwim-complete.el]:
  * create a standard completion API that is simple but uses
    a sophisticated backed.

This feature is now complete. The dwim-complete interface programmatically
is flexible and powerful. The user interface of helm is fully adapted.

emacs-lisp.el and template.el are included as examples of use.