<buffer-ring> half fixed the problem with can't switch to buffer.
when changing major modes which happens on a reload it looks like the
whole damn buffer object dissapears causing dangling buffer id's in
the ring.

I tried various ways to deal with this including a significant rewrite
to use buffer name strings instead of buffer objects however the problem
persists.

the fix for now is to sweep the buffer ring for dead refs every time
it is used in the interface. this works for now.

a permanent solution may involve hooking major-mode-change-hook or
whatever it is called to save/restore and see if the problem is the
mode change or if the buffer is completely destroyed and re-created.